Diagrams.Direction

Description

Type for representing directions, polymorphic in vector space

Documentation

data Direction v n

A vector is described by a Direction and a magnitude. So we can think of a Direction as a vector that has forgotten its magnitude. Directions can be used with fromDirection and the lenses provided by its instances.

_Dir :: Iso' (Direction v n) (v n)

_Dir is provided to allow efficient implementations of functions in particular vector-spaces, but should be used with care as it exposes too much information.

direction :: v n -> Direction v n

direction v is the direction in which v points. Returns an unspecified value when given the zero vector as input.

dir :: v n -> Direction v n

Synonym for direction.

fromDirection :: (Metric v, Floating n) => Direction v n -> v n

fromDirection d is the unit vector in the direction d.

fromDir :: (Metric v, Floating n) => Direction v n -> v n

Synonym for fromDirection.

angleBetweenDirs :: (Metric v, Floating n) => Direction v n -> Direction v n -> Angle n

compute the positive angle between the two directions in their common plane

dirBetween :: (Additive v, Num n) => Point v n -> Point v n -> Direction v n

dirBetween p q returns the directions from p to q