Diagrams.Points

Description

Points in space. For more tools for working with points and vectors, see Linear.Affine.

Points

newtype Point f a :: (* -> *) -> * -> *

A handy wrapper to help distinguish points from vectors at the type level

Constructors

P (f a)

origin :: (Additive f, Num a) => Point f a

Vector spaces have origins.

(*.) :: (Functor v, Num n) => n -> Point v n -> Point v n

Scale a point by a scalar. Specialized version of '(*^)'.

Point-related utilities

centroid :: (Additive v, Fractional n) => [Point v n] -> Point v n

The centroid of a set of n points is their sum divided by n.

pointDiagram :: (Metric v, Fractional n) => Point v n -> QDiagram b v n m

Create a "point diagram", which has no content, no trace, an empty query, and a point envelope.

_Point :: (Profunctor p, Functor f) => p (f a) (f (f a)) -> p (Point f a) (f (Point f a))

lensP :: Functor f => (g a -> f (g a)) -> Point g a -> f (Point g a)