Diagrams.Solve

Description

Exact solving of low-degree (n <= 4) polynomials.

Documentation

quadForm :: (Floating d, Ord d) => d -> d -> d -> [d]

The quadratic formula.

cubForm :: (Floating d, Ord d) => d -> d -> d -> d -> [d]

Solve the cubic equation ax^3 + bx^2 + cx + d = 0, returning a list of all real roots within 1e-10 tolerance (although currently it's closer to 1e-5)

quartForm :: (Floating d, Ord d) => d -> d -> d -> d -> d -> [d]

Solve the quartic equation c4 x^4 + c3 x^3 + c2 x^2 + c1 x + c0 = 0, returning a list of all real roots within 1e-10 tolerance (although currently it's closer to 1e-5)

cubForm' :: (Floating d, Ord d) => d -> d -> d -> d -> d -> [d]

Solve the cubic equation ax^3 + bx^2 + cx + d = 0, returning a list of all real roots. First argument is tolerance.

quartForm' :: (Floating d, Ord d) => d -> d -> d -> d -> d -> d -> [d]

Solve the quartic equation c4 x^4 + c3 x^3 + c2 x^2 + c1 x + c0 = 0, returning a list of all real roots. First argument is tolerance.