Diagrams.ThreeD.Shapes

Description

Various three-dimensional shapes.

Documentation

data Ellipsoid

Constructors

Ellipsoid T3

sphere :: (Backend b R3, Renderable Ellipsoid b) => Diagram b R3

A sphere of radius 1 with its center at the origin.

data Box

Constructors

Box T3

cube :: (Backend b R3, Renderable Box b) => Diagram b R3

A cube with side length 1, in the positive octant, with one vertex at the origin.

data Frustum

Constructors

Frustum Double Double T3

frustum :: (Backend b R3, Renderable Frustum b) => Double -> Double -> Diagram b R3

A frustum of a right circular cone. It has height 1 oriented along the positive z axis, and radii r0 and r1 at Z=0 and Z=1. cone and cylinder are special cases.

cone :: (Backend b R3, Renderable Frustum b) => Diagram b R3

A cone with its base centered on the origin, with radius 1 at the base, height 1, and it's apex on the positive Z axis.

cylinder :: (Backend b R3, Renderable Frustum b) => Diagram b R3

A circular cylinder of radius 1 with one end cap centered on the origin, and extending to Z=1.