Diagrams.ThreeD.Vector

Description

Three-dimensional vectors.

Special 2D vectors

unitX :: R3

The unit vector in the positive X direction.

unitY :: R3

The unit vector in the positive Y direction.

unitZ :: R3

The unit vector in the positive Z direction.

unit_X :: R3

The unit vector in the negative X direction.

unit_Y :: R3

The unit vector in the negative Y direction.

unit_Z :: R3

The unit vector in the negative Z direction.

Converting between vectors and angles

direction :: Direction d => R3 -> d

direction v is the direction in which v points. Returns an unspecified value when given the zero vector as input.

fromDirection :: Direction d => d -> R3

fromDirection d is the unit vector in the direction d.

angleBetween :: (Angle a, Num a, Ord a) => R3 -> R3 -> a

compute the positive angle between the two vectors in their common plane

angleBetweenDirs :: (Direction d, Angle a, Num a, Ord a) => d -> d -> a

compute the positive angle between the two vectors in their common plane