Diagrams.Trace

Description

"Traces", aka embedded raytracers, for finding points on the edge of a diagram. See Diagrams.Core.Trace for internal implementation details.

Types

data Trace v

Every diagram comes equipped with a trace. Intuitively, the trace for a diagram is like a raytracer: given a line (represented as a base point and a direction), the trace computes the distance from the base point along the line to the first intersection with the diagram. The distance can be negative if the intersection is in the opposite direction from the base point, or infinite if the ray never intersects the diagram. Note: to obtain the distance to the furthest intersection instead of the closest, just negate the direction vector and then negate the result.

Note that the output should actually be interpreted not as an absolute distance, but as a multiplier relative to the input vector. That is, if the input vector is v and the returned scalar is s, the distance from the base point to the intersection is given by s * magnitude v.

class (Ord (Scalar (V a)), VectorSpace (V a)) => Traced a

Traced abstracts over things which have a trace.

Diagram traces

trace :: (Ord (Scalar v), VectorSpace v, HasLinearMap v) => QDiagram b v m -> Trace v

Get the trace of a diagram.

setTrace :: (OrderedField (Scalar v), InnerSpace v, HasLinearMap v, Semigroup m) => Trace v -> QDiagram b v m -> QDiagram b v m

Replace the trace of a diagram.

withTrace :: (HasLinearMap (V a), Traced a, OrderedField (Scalar (V a)), InnerSpace (V a), Monoid' m) => a -> QDiagram b (V a) m -> QDiagram b (V a) m

Use the trace from some object as the trace for a diagram, in place of the diagram's default trace.

Querying traces

traceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

Compute the vector from the given point to the boundary of the given object in the given direction, or Nothing if there is no intersection.

traceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

Given a base point and direction, compute the closest point on the boundary of the given object, or Nothing if there is no intersection in the given direction.

maxTraceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

Like traceV, but computes a vector to the *furthest* point on the boundary instead of the closest.

maxTraceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

Like traceP, but computes the *furthest* point on the boundary instead of the closest.

Subdiagram traces

boundaryFrom :: (HasLinearMap v, Ord (Scalar v)) => Subdiagram b v m -> v -> Point v

Compute the furthest point on the boundary of a subdiagram, beginning from the location (local origin) of the subdiagram and moving in the direction of the given vector. If there is no such point, the origin is returned; see also boundaryFromMay.

boundaryFromMay :: (HasLinearMap v, Ord (Scalar v)) => Subdiagram b v m -> v -> Maybe (Point v)

Compute the furthest point on the boundary of a subdiagram, beginning from the location (local origin) of the subdiagram and moving in the direction of the given vector, or Nothing if there is no such point.