Diagrams.Transform.Matrix

Description

Functions for converting between Transformations and matricies.

Documentation

mkMat :: (HasBasis v, Num n) => Transformation v n -> v (v n)

Build a matrix from a Transformation, ignoring the translation.

mkMatHomo :: Num n => Transformation V3 n -> M44 n

Build a 3D transformation matrix in homogeneous coordinates from a 'Transformation V3'.

fromMat22 :: (Epsilon n, Floating n) => M22 n -> V2 n -> Maybe (T2 n)

Make a 2D transformation from a 2x2 transform matrix and a translation vector. If the matrix is not invertible, Nothing is returned.

fromMat33 :: (Epsilon n, Floating n) => M33 n -> V3 n -> Maybe (T3 n)

Make a 3D transformation from a 3x3 transform matrix and a translation vector. If the matrix is not invertible, Nothing is returned.

fromMatWithInv

Arguments

:: (Additive v, Distributive v, Foldable v, Num n)

=> v (v n)
matrix

-> v (v n)
inverse

-> v n
translation

-> Transformation v n

Build a transform with a maxtrix along with its inverse.

mat22 :: (Epsilon n, Floating n) => Prism' (M22 n, V2 n) (T2 n)

Prism onto a 2D transformation from a 2x2 transform matrix and translation vector.

mat33 :: (Epsilon n, Floating n) => Prism' (M33 n, V3 n) (T3 n)

Prism onto a 2D transformation from a 2x2 transform matrix and translation vector.