Diagrams.TwoD.Arc

Description

Two-dimensional arcs, approximated by cubic bezier curves.

Documentation

arc :: (Angle a, PathLike p, V p ~ R2) => a -> a -> p

Given a start angle s and an end angle e, arc s e is the path of a radius one arc counterclockwise between the two angles.

arcT :: Angle a => a -> a -> Trail R2

bezierFromSweep :: Rad -> [Segment R2]

bezierFromSweep s constructs a series of Cubic segments that start in the positive y direction and sweep counter clockwise through s radians. If s is negative, it will start in the negative y direction and sweep clockwise. When s is less than 0.0001 the empty list results. If the sweep is greater than tau then it is truncated to tau.

wedge :: (Angle a, PathLike p, V p ~ R2) => Double -> a -> a -> p

Create a circular wedge of the given radius, beginning at the first angle and extending counterclockwise to the second.