Diagrams.TwoD.Image

Description

Importing external images into diagrams.

Documentation

data Image

An external image primitive, representing an image the backend should import from another file when rendering.

Constructors

Image

Fields

_imgFile :: FilePath

_imgSize :: SizeSpec2D

_imgTransf :: T2

imgFile :: Lens' Image FilePath

imgSize :: Lens' Image SizeSpec2D

imgTransf :: Lens' Image T2

image :: Renderable Image b => FilePath -> Double -> Double -> Diagram b R2

Take an external image from the specified file and turn it into a diagram with the specified width and height, centered at the origin. Note that the image's aspect ratio will be preserved; if the specified width and height have a different ratio than the image's aspect ratio, there will be extra space in one dimension.