Diagrams.Attributes

Description

Diagrams may have attributes which affect the way they are rendered. This module defines some common attributes; particular backends may also define more backend-specific attributes.

Every attribute type must have a semigroup structure, that is, an associative binary operation for combining two attributes into one. Unless otherwise noted, all the attributes defined here use the Last structure, that is, combining two attributes simply keeps the second one and throws away the first. This means that child attributes always override parent attributes.

Color

Diagrams outsources all things color-related to Russell O'Connor's very nice colour package (http://hackage.haskell.org/package/colour). For starters, it provides a large collection of standard color names. However, it also provides a rich set of combinators for combining and manipulating colors; see its documentation for more information.

class Color c where

The Color type class encompasses color representations which can be used by the Diagrams library. Instances are provided for both the Colour and AlphaColour types from the Data.Colour library.

Methods

toAlphaColour :: c -> AlphaColour Double

Convert a color to its standard representation, AlphaColour.

fromAlphaColour :: AlphaColour Double -> c

Convert from an AlphaColour Double. Note that this direction may lose some information. For example, the instance for Colour drops the alpha channel.

data SomeColor

An existential wrapper for instances of the Color class.

Constructors

forall c . Color c => SomeColor c

someToAlpha :: SomeColor -> AlphaColour Double

Line color

data LineColor

The color with which lines (strokes) are drawn. Note that child colors always override parent colors; that is, lineColor c1 . lineColor c2 $ d is equivalent to lineColor c2 $ d. More precisely, the semigroup structure on line color attributes is that of Last.

getLineColor :: LineColor -> SomeColor

mkLineColor :: Color c => c -> LineColor

styleLineColor :: (Color c, Color c') => Setter (Style v) (Style v) c c'

lineColor :: (Color c, HasStyle a) => c -> a -> a

Set the line (stroke) color. This function is polymorphic in the color type (so it can be used with either Colour or AlphaColour), but this can sometimes create problems for type inference, so the lc and lcA variants are provided with more concrete types.

lineColorA :: HasStyle a => LineColor -> a -> a

Apply a lineColor attribute.

lc :: HasStyle a => Colour Double -> a -> a

A synonym for lineColor, specialized to Colour Double (i.e. opaque colors).

lcA :: HasStyle a => AlphaColour Double -> a -> a

A synonym for lineColor, specialized to AlphaColour Double (i.e. colors with transparency).

Fill color

data FillColor

The color with which shapes are filled. Note that child colors always override parent colors; that is, fillColor c1 . fillColor c2 $ d is equivalent to lineColor c2 $ d. More precisely, the semigroup structure on fill color attributes is that of Last.

getFillColor :: FillColor -> SomeColor

mkFillColor :: Color c => c -> FillColor

styleFillColor :: (Color c, Color c') => Setter (Style v) (Style v) c c'

recommendFillColor :: (Color c, HasStyle a) => c -> a -> a

Set a "recommended" fill color, to be used only if no explicit calls to fillColor (or fc, or fcA) are used.

fillColor :: (Color c, HasStyle a) => c -> a -> a

Set the fill color. This function is polymorphic in the color type (so it can be used with either Colour or AlphaColour), but this can sometimes create problems for type inference, so the fc and fcA variants are provided with more concrete types.

fc :: HasStyle a => Colour Double -> a -> a

A synonym for fillColor, specialized to Colour Double (i.e. opaque colors).

fcA :: HasStyle a => AlphaColour Double -> a -> a

A synonym for fillColor, specialized to AlphaColour Double (i.e. colors with transparency).

Opacity

data Opacity

Although the individual colors in a diagram can have transparency, the opacity/transparency of a diagram as a whole can be specified with the Opacity attribute. The opacity is a value between 1 (completely opaque, the default) and 0 (completely transparent). Opacity is multiplicative, that is, opacity o1 . opacity o2 === opacity (o1 * o2). In other words, for example, opacity 0.8 means "decrease this diagram's opacity to 80% of its previous opacity".

getOpacity :: Opacity -> Double

opacity :: HasStyle a => Double -> a -> a

Multiply the opacity (see Opacity) by the given value. For example, opacity 0.8 means "decrease this diagram's opacity to 80% of its previous opacity".

Converting colors

colorToSRGBA :: Color c => c -> (Double, Double, Double, Double)

Convert to sRGBA.

colorToRGBA :: Color c => c -> (Double, Double, Double, Double)

Deprecated: Renamed to colorToSRGBA.

Convert to sRGBA.

Lines

Width

data LineWidth

The width of lines. By default, the line width is measured with respect to the final coordinate system of a rendered diagram, as opposed to the local coordinate systems in effect at the time the line width was set for various subdiagrams. This is so that it is easy to combine a variety of shapes (some created by scaling) and have them all drawn using a consistent line width. However, sometimes it is desirable for scaling to affect line width; the freeze operation is provided for this purpose. The line width of frozen diagrams is affected by transformations.

Line widths specified on child nodes always override line widths specified at parent nodes.

getLineWidth :: LineWidth -> Double

lineWidth :: HasStyle a => Double -> a -> a

Set the line (stroke) width.

lineWidthA :: HasStyle a => LineWidth -> a -> a

Apply a LineWidth attribute.

lw :: HasStyle a => Double -> a -> a

A convenient synonym for lineWidth.

Cap style

data LineCap

What sort of shape should be placed at the endpoints of lines?

Constructors

LineCapButt
Lines end precisely at their endpoints.

LineCapRound
Lines are capped with semicircles centered on endpoints.

LineCapSquare
Lines are capped with a squares centered on endpoints.

data LineCapA

getLineCap :: LineCapA -> LineCap

lineCap :: HasStyle a => LineCap -> a -> a

Set the line end cap attribute.

Join style

data LineJoin

How should the join points between line segments be drawn?

Constructors

LineJoinMiter
Use a "miter" shape (whatever that is).

LineJoinRound
Use rounded join points.

LineJoinBevel
Use a "bevel" shape (whatever that is). Are these... carpentry terms?

data LineJoinA

getLineJoin :: LineJoinA -> LineJoin

lineJoin :: HasStyle a => LineJoin -> a -> a

Set the segment join style.

Miter limit

newtype LineMiterLimit

Miter limit attribute affecting the LineJoinMiter joins. For some backends this value may have additional effects.

Constructors

LineMiterLimit (Last Double)

getLineMiterLimit :: LineMiterLimit -> Double

lineMiterLimit :: HasStyle a => Double -> a -> a

Set the miter limit for joins with LineJoinMiter.

lineMiterLimitA :: HasStyle a => LineMiterLimit -> a -> a

Apply a LineMiterLimit attribute.

Dashing

data Dashing

Create lines that are dashing... er, dashed.

Constructors

Dashing [Double] Double

data DashingA

getDashing :: DashingA -> Dashing

dashing

Arguments

:: HasStyle a

=> [Double]
A list specifying alternate lengths of on and off portions of the stroke. The empty list indicates no dashing.

-> Double
An offset into the dash pattern at which the stroke should start.

-> a

-> a

Set the line dashing style.

Compilation utilities

splitFills :: forall b v a. Typeable v => RTree b v a -> RTree b v a

Push fill attributes down until they are at the root of subtrees containing only loops. This makes life much easier for backends, which typically have a semantics where fill attributes are applied to linesnon-closed paths as well as loopsclosed paths, whereas in the semantics of diagrams, fill attributes only apply to loops.