Diagrams.Attributes.Compile

Description

XXX

Documentation

class (AttributeClass (AttrType code), Typeable (PrimType code)) => SplitAttribute code where

Associated Types

type AttrType code :: Type

type PrimType code :: Type

Methods

primOK :: code -> PrimType code -> Bool

splitAttr :: forall code b v n a. SplitAttribute code => code -> RTree b v n a -> RTree b v n a

Push certain attributes down until they are at the roots of trees containing only "safe" nodes. In particular this is used to push fill attributes down until they are over only loops; see splitFills.