Diagrams.Coordinates

Description

Nice syntax for constructing and pattern-matching on literal points and vectors.

Documentation

data a :& b

A pair of values, with a convenient infix (left-associative) data constructor.

Constructors

a :& b

class Coordinates c where

Types which are instances of the Coordinates class can be constructed using ^& (for example, a three-dimensional vector could be constructed by 1 ^& 6 ^& 3), and deconstructed using coords. A common pattern is to use coords in conjunction with the ViewPatterns extension, like so:

 foo :: Vector3 -> ...
 foo (coords -> x :& y :& z) = ...

Associated Types

type FinalCoord c :: *

The type of the final coordinate.

type PrevDim c :: *

The type of everything other than the final coordinate.

type Decomposition c :: *

Decomposition of c into applications of :&.

Methods

(^&) :: PrevDim c -> FinalCoord c -> c

Construct a value of type c by providing something of one less dimension (which is perhaps itself recursively constructed using (^&)) and a final coordinate. For example,

 2 ^& 3 :: P2
 3 ^& 5 ^& 6 :: R3

Note that ^& is left-associative.

pr :: PrevDim c -> FinalCoord c -> c

Prefix synonym for ^&. pr stands for pair of PrevDim, FinalCoord

coords :: c -> Decomposition c

Decompose a value of type c into its constituent coordinates, stored in a nested (:&) structure.

Lenses for particular axes

class HasX t where

The class of types with at least one coordinate, called _x.

Methods

_x :: Lens' t Double

class HasY t where

The class of types with at least two coordinates, the second called _y.

Methods

_y :: Lens' t Double

class HasZ t where

The class of types with at least three coordinates, the third called _z.

Methods

_z :: Lens' t Double