Diagrams.CubicSpline.Internal

Description

A cubic spline is a smooth, connected sequence of cubic curves passing through a given sequence of points. This module implements a straightforward spline generation algorithm based on solving tridiagonal systems of linear equations.

Solving for spline coefficents

solveCubicSplineDerivatives :: Fractional a => [a] -> [a]

Use the tri-diagonal solver with the appropriate parameters for an open cubic spline.

solveCubicSplineDerivativesClosed :: Fractional a => [a] -> [a]

Use the cyclic-tri-diagonal solver with the appropriate parameters for a closed cubic spline.

solveCubicSplineCoefficients :: Fractional a => Bool -> [a] -> [[a]]

Use the cyclic-tri-diagonal solver with the appropriate parameters for a closed cubic spline.