Diagrams.LinearMap

Description

Linear maps. Unlike Transformations these are not restricted to the same space. In practice these are used for projections in Diagrams.ThreeD.Projection. Unless you want to work with projections you're probably better off using Transform.

Currently only path-like things can be projected. In the future we hope to support projecting diagrams.

Linear maps

newtype LinearMap v u n

Type for holding linear maps. Note that these are not affine transforms so attemping apply a translation with LinearMap will likely produce incorrect results.

Constructors

LinearMap

Fields

lapply :: v n -> u n

class LinearMappable a b where

Class of things that have vectors that can be mapped over.

Methods

vmap :: (Vn a -> Vn b) -> a -> b

Apply a linear map to an object. If the map is not linear, behaviour will likely be wrong.

Applying linear maps

linmap :: (InSpace v n a, LinearMappable a b, N b ~ n) => LinearMap v (V b) n -> a -> b

Apply a linear map.

Affine maps

data AffineMap v u n

Affine linear maps. Unlike Transformation these do not have to be invertible so we can map between spaces.

Constructors

AffineMap (LinearMap v u n) (u n)

class (LinearMappable a b, N a ~ N b) => AffineMappable a b where

Minimal complete definition

Nothing

Methods

amap :: (Additive (V a), Foldable (V a), Additive (V b), Num (N b)) => AffineMap (V a) (V b) (N b) -> a -> b

Affine map over an object. Has a default implimentation of only applying the linear map

Constructing affine maps

mkAffineMap :: (v n -> u n) -> u n -> AffineMap v u n

Make an affine map from a linear function and a translation.

toAffineMap :: Transformation v n -> AffineMap v v n