Diagrams.Parametric.Adjust

Description

Tools for adjusting the length of parametric objects such as segments and trails.

Documentation

adjust :: (DomainBounds a, Sectionable a, HasArcLength a, Fractional (Scalar (V a))) => a -> AdjustOpts (V a) -> a

Adjust the length of a parametric object such as a segment or trail. The second parameter is an option record which controls how the adjustment should be performed; see AdjustOpts.

data AdjustOpts v

How should a segment, trail, or path be adjusted?

adjMethod :: Lens' (AdjustOpts v) (AdjustMethod v)

Which method should be used for adjusting?

adjSide :: Lens' (AdjustOpts v) AdjustSide

Which end(s) of the object should be adjusted?

adjEps :: Lens' (AdjustOpts v) (Scalar v)

Tolerance to use when doing adjustment.

data AdjustMethod v

What method should be used for adjusting a segment, trail, or path?

Constructors

ByParam (Scalar v)
Extend by the given parameter value (use a negative parameter to shrink)

ByAbsolute (Scalar v)
Extend by the given arc length (use a negative length to shrink)

ToAbsolute (Scalar v)
Extend or shrink to the given arc length

data AdjustSide

Which side of a segment, trail, or path should be adjusted?

Constructors

Start
Adjust only the beginning

End
Adjust only the end

Both
Adjust both sides equally