Diagrams.Path

Description

This module defines trails (translationally invariant sequences of linear or cubic Bézier segments) and paths (collections of concretely located trails). Trails and paths can be used for drawing shapes, laying out other diagrams, clipping, and other things.

Constructing path-like things

class (Monoid' p, VectorSpace (V p)) => PathLike p where

Type class for path-like things, which must be monoids. Instances include Trails, Paths, and two-dimensional Diagrams.

Methods

pathLike

Arguments

:: Point (V p)
The starting point of the path. Some path-like things (e.g. Trails) may ignore this.

-> Bool
Should the path be closed?

-> [Segment (V p)]
Segments of the path.

-> p

fromSegments :: PathLike p => [Segment (V p)] -> p

Construct an open path-like thing with the origin as a starting point.

fromOffsets :: PathLike p => [V p] -> p

Construct an open path-like thing of linear segments from a list of offsets. The starting point is the origin.

fromVertices :: PathLike p => [Point (V p)] -> p

Construct a path-like thing of linear segments from a list of vertices, with the first vertex as the starting point.

segmentsFromVertices :: AdditiveGroup v => [Point v] -> [Segment v]

Construct a list of linear segments from a list of vertices. The input list must contain at least two points to generate a non-empty list of segments.

pathLikeFromTrail :: PathLike p => Trail (V p) -> p

Convert a trail to any path-like thing. pathLikeFromTrail is the identity on trails.

Closeable things

class PathLike p => Closeable p where

Path-like things that can be "open" or "closed".

Methods

open :: p -> p

"Open" a path-like thing.

close :: p -> p

"Close" a path-like thing, by implicitly connecting the endpoint(s) back to the starting point(s).

Trails

data Trail v

A trail is a sequence of segments placed end-to-end. Trails are thus translationally invariant, and form a monoid under concatenation. Trails can also be open (the default) or closed (the final point in a closed trail is implicitly connected back to the starting point).

Constructors

Trail

Fields

trailSegments :: [Segment v]

isClosed :: Bool

Computing with trails

trailSegments' :: AdditiveGroup v => Trail v -> [Segment v]

trailSegments' is like trailSegments, but explicitly includes the implicit closing segment at the end of the list for closed trails.

trailOffsets :: Trail v -> [v]

Extract the offsets of the segments of a trail.

trailOffset :: AdditiveGroup v => Trail v -> v

Compute the offset from the start of a trail to the end.

trailVertices :: AdditiveGroup v => Point v -> Trail v -> [Point v]

Extract the vertices of a trail, given a concrete location at which to place the first vertex.

reverseTrail :: AdditiveGroup v => Trail v -> Trail v

Reverse a trail's direction of travel.

addClosingSegment :: AdditiveGroup v => Trail v -> Trail v

If the trail is closed, this adds in the closing segment. Otherwise, the trail is returned unmodified.

fixTrail :: AdditiveGroup v => Point v -> Trail v -> [FixedSegment v]

Convert a starting point and a trail into a list of fixed segments.

Paths

newtype Path v

A path is a (possibly empty) list of trails, with each trail paired with an absolute starting point. Hence, paths are not translationally invariant, and form a monoid under superposition.

Constructors

Path

Fields

pathTrails :: [(Point v, Trail v)]

Constructing paths from trails

pathFromTrail :: AdditiveGroup v => Trail v -> Path v

Convert a trail to a path beginning at the origin.

pathFromTrailAt :: Trail v -> Point v -> Path v

Convert a trail to a path with a particular starting point.

Computing with paths

pathVertices :: AdditiveGroup v => Path v -> [[Point v]]

Extract the vertices of a path.

pathOffsets :: AdditiveGroup v => Path v -> [v]

Compute the total offset of each trail comprising a path.

pathCentroid :: (VectorSpace v, Fractional (Scalar v)) => Path v -> Point v

Compute the centroid of a path (i.e. the average of its vertices).

expandPath :: (HasLinearMap v, VectorSpace v, Fractional (Scalar v), Eq (Scalar v)) => Scalar v -> Path v -> Path v

Scale a path using its centroid (see pathCentroid) as the base point for the scale.

reversePath :: AdditiveGroup v => Path v -> Path v

Reverse the direction of all the component trails of a path.

fixPath :: AdditiveGroup v => Path v -> [[FixedSegment v]]

Convert a path into a list of lists of FixedSegments.

Miscellaneous

explodeTrail :: (VectorSpace (V p), PathLike p) => Point (V p) -> Trail (V p) -> [p]

Given a starting point, "explode" a trail by turning each segment (including the implicit closing segment, if the trail is closed) into its own separate path. Useful for (say) applying a different style to each segment.

explodePath :: (VectorSpace (V p), PathLike p) => Path (V p) -> [[p]]

"Explode" a path by exploding every component trail (see explodeTrail).

(~~) :: PathLike p => Point (V p) -> Point (V p) -> p

Create a single-segment path between two given points.