Diagrams.Query

Description

A query is a function that maps points in a vector space to values in some monoid. Queries naturally form a monoid, with two queries being combined pointwise.

Queries

newtype Query v n m :: (* -> *) -> * -> * -> *

A query is a function that maps points in a vector space to values in some monoid. Queries naturally form a monoid, with two queries being combined pointwise.

The idea for annotating diagrams with monoidal queries came from the graphics-drawingcombinators package, http://hackage.haskell.org/package/graphics-drawingcombinators.

Constructors

Query

Fields

runQuery :: Point v n -> m

class HasQuery t m | t -> m where

Types which can answer a Query about points inside the geometric object.

If t and m are both a Semigroups, getQuery should satisfy

getQuery (t1 <> t2) = getQuery t1 <> getQuery t2

Minimal complete definition

getQuery

Methods

getQuery :: t -> Query (V t) (N t) m

Extract the query of an object.

sample :: HasQuery t m => t -> Point (V t) (N t) -> m

Sample a diagram's query function at a given point.

sample :: QDiagram b v n m -> Point v n -> m
sample :: Query v n m      -> Point v n -> m
sample :: BoundingBox v n  -> Point v n -> Any
sample :: Path V2 Double   -> Point v n -> Crossings

inquire :: HasQuery t Any => t -> Point (V t) (N t) -> Bool

Test if a point is not equal to mempty.

inquire :: QDiagram b v n Any -> Point v n -> Bool
inquire :: Query v n Any      -> Point v n -> Bool
inquire :: BoundingBox v n  -> Point v n -> Bool

Queries on diagrams

query :: Monoid m => QDiagram b v n m -> Query v n m

Get the query function associated with a diagram.

value :: Monoid m => m -> QDiagram b v n Any -> QDiagram b v n m

Set the query value for True points in a diagram (i.e. points "inquire" the diagram); False points will be set to mempty.

resetValue :: (Eq m, Monoid m) => QDiagram b v n m -> QDiagram b v n Any

Reset the query values of a diagram to True/False: any values equal to mempty are set to False; any other values are set to True.

clearValue :: QDiagram b v n m -> QDiagram b v n Any

Set all the query values of a diagram to False.