Diagrams.Query

Description

A query is a function that maps points in a vector space to values in some monoid. Queries naturally form a monoid, with two queries being combined pointwise.

Documentation

newtype Query v m

Constructors

Query

Fields

runQuery :: Point v -> m

query :: Monoid m => QDiagram b v m -> Query v m

sample :: Monoid m => QDiagram b v m -> Point v -> m

value :: Monoid m => m -> QDiagram b v Any -> QDiagram b v m

resetValue :: (Eq m, Monoid m) => QDiagram b v m -> QDiagram b v Any

clearValue :: QDiagram b v m -> QDiagram b v Any