Diagrams.ThreeD.Light

Description

Types to specify lighting for 3D rendering.

Documentation

data PointLight n

A PointLight radiates uniformly in all directions from a given point.

Constructors

PointLight (Point V3 n) (Colour Double)

data ParallelLight n

A ParallelLight casts parallel rays in the specified direction, from some distant location outside the scene.

Constructors

ParallelLight (V3 n) (Colour Double)

pointLight

Arguments

:: (Typeable n, Num n, Ord n, Renderable (PointLight n) b)

=> Colour Double
The color of the light

-> QDiagram b V3 n Any

Construct a Diagram with a single PointLight at the origin, which takes up no space.

parallelLight

Arguments

:: (Typeable n, OrderedField n, Renderable (ParallelLight n) b)

=> Direction V3 n
The direction in which the light travels.

-> Colour Double
The color of the light.

-> QDiagram b V3 n Any

Construct a Diagram with a single ParallelLight, which takes up no space.