Diagrams.ThreeD.Shapes

Description

Various three-dimensional shapes.

Documentation

data Ellipsoid n

Constructors

Ellipsoid (Transformation V3 n)

sphere :: (Typeable n, OrderedField n, Renderable (Ellipsoid n) b) => QDiagram b V3 n Any

A sphere of radius 1 with its center at the origin.

data Box n

Constructors

Box (Transformation V3 n)

cube :: (Typeable n, OrderedField n, Renderable (Box n) b) => QDiagram b V3 n Any

A cube with side length 1, in the positive octant, with one vertex at the origin.

data Frustum n

Constructors

Frustum n n (Transformation V3 n)

frustum :: (TypeableFloat n, Renderable (Frustum n) b) => n -> n -> QDiagram b V3 n Any

A frustum of a right circular cone. It has height 1 oriented along the positive z axis, and radii r0 and r1 at Z=0 and Z=1. cone and cylinder are special cases.

cone :: (TypeableFloat n, Renderable (Frustum n) b) => QDiagram b V3 n Any

A cone with its base centered on the origin, with radius 1 at the base, height 1, and it's apex on the positive Z axis.

cylinder :: (TypeableFloat n, Renderable (Frustum n) b) => QDiagram b V3 n Any

A circular cylinder of radius 1 with one end cap centered on the origin, and extending to Z=1.