Diagrams.ThreeD.Shapes

Description

Various three-dimensional shapes.

Skinned class

class Skinned t where

Types which can be rendered as 3D Diagrams.

Minimal complete definition

skin

Methods

skin :: (Renderable t b, N t ~ n, TypeableFloat n) => t -> QDiagram b V3 n Any

Basic 3D shapes

data Ellipsoid n

Constructors

Ellipsoid (Transformation V3 n)

sphere :: Num n => Ellipsoid n

A sphere of radius 1 with its center at the origin.

data Box n

Constructors

Box (Transformation V3 n)

cube :: Num n => Box n

A cube with side length 1, in the positive octant, with one vertex at the origin.

data Frustum n

Constructors

Frustum n n (Transformation V3 n)

frustum :: Num n => n -> n -> Frustum n

A frustum of a right circular cone. It has height 1 oriented along the positive z axis, and radii r0 and r1 at Z=0 and Z=1. cone and cylinder are special cases.

cone :: Num n => Frustum n

A cone with its base centered on the origin, with radius 1 at the base, height 1, and it's apex on the positive Z axis.

cylinder :: Num n => Frustum n

A circular cylinder of radius 1 with one end cap centered on the origin, and extending to Z=1.

Constructive solid geometry

data CSG n

A tree of Constructive Solid Geometry operations and the primitives that can be used in them.

Constructors

CsgEllipsoid (Ellipsoid n)

CsgBox (Box n)

CsgFrustum (Frustum n)

CsgUnion [CSG n]

CsgIntersection [CSG n]

CsgDifference (CSG n) (CSG n)

union :: (CsgPrim a, CsgPrim b) => a n -> b n -> CSG n

intersection :: (CsgPrim a, CsgPrim b) => a n -> b n -> CSG n

difference :: (CsgPrim a, CsgPrim b) => a n -> b n -> CSG n