Diagrams.ThreeD.Size

Description

Utilities for working with sizes of three-dimensional objects.

Computing sizes

extentX :: (InSpace v n a, R1 v, Enveloped a) => a -> Maybe (n, n)

Compute the absolute x-coordinate range of an enveloped object in the form (lo,hi). Return Nothing for objects with an empty envelope.

Note this is just extent unitX.

extentY :: (InSpace v n a, R2 v, Enveloped a) => a -> Maybe (n, n)

Compute the absolute y-coordinate range of an enveloped object in the form (lo,hi). Return Nothing for objects with an empty envelope.

extentZ :: (InSpace v n a, R3 v, Enveloped a) => a -> Maybe (n, n)

Compute the absolute z-coordinate range of an enveloped object in the form (lo,hi). Return Nothing for objects with an empty envelope.

Specifying sizes

mkSizeSpec3D :: Num n => Maybe n -> Maybe n -> Maybe n -> SizeSpec V3 n

Make a SizeSpec from possibly-specified width and height.

dims3D :: n -> n -> n -> SizeSpec V3 n

Make a SizeSpec from a width and height.