Diagrams.ThreeD.Types

Description

Basic types for three-dimensional Euclidean space.

3D Euclidean space

data R3

The three-dimensional Euclidean vector space R^3.

r3 :: (Double, Double, Double) -> R3

Construct a 3D vector from a triple of components.

unr3 :: R3 -> (Double, Double, Double)

Convert a 3D vector back into a triple of components.

mkR3 :: Double -> Double -> Double -> R3

Curried version of r3.

type P3 = Point R3

Points in R^3.

p3 :: (Double, Double, Double) -> P3

Construct a 3D point from a triple of coordinates.

unp3 :: P3 -> (Double, Double, Double)

Convert a 3D point back into a triple of coordinates.

mkP3 :: Double -> Double -> Double -> P3

Curried version of r3.

type T3 = Transformation R3

Transformations in R^3.

r3Iso :: Iso' R3 (Double, Double, Double)

p3Iso :: Iso' P3 (Double, Double, Double)

Two-dimensional angles

These are defined in Diagrams.TwoD.Types but reëxported here for convenience.

class Num a => Angle a where

Type class for types that measure angles.

Methods

toTurn :: a -> Turn

Convert to a turn, i.e. a fraction of a circle.

fromTurn :: Turn -> a

Convert from a turn, i.e. a fraction of a circle.

newtype Turn

Newtype wrapper used to represent angles as fractions of a circle. For example, 1/3 turn = tau/3 radians = 120 degrees.

Constructors

Turn Double

asTurn :: Turn -> Turn

The identity function with a restricted type, for conveniently declaring that some value should have type Turn. For example, rotation . asTurn . fromRational constructs a rotation from a rational value considered as a Turn. Without asTurn, the angle type would be ambiguous.

type CircleFrac = Turn

Deprecated synonym for Turn, retained for backwards compatibility.

newtype Rad

Newtype wrapper for representing angles in radians.

Constructors

Rad Double

asRad :: Rad -> Rad

The identity function with a restricted type, for conveniently declaring that some value should have type Rad. For example, rotation . asRad . fromRational constructs a rotation from a rational value considered as a value in radians. Without asRad, the angle type would be ambiguous.

newtype Deg

Newtype wrapper for representing angles in degrees.

Constructors

Deg Double

asDeg :: Deg -> Deg

The identity function with a restricted type, for conveniently declaring that some value should have type Deg. For example, rotation . asDeg . fromIntegral constructs a rotation from an integral value considered as a value in degrees. Without asDeg, the angle type would be ambiguous.

fullTurn :: Angle a => a

An angle representing one full turn.

convertAngle :: (Angle a, Angle b) => a -> b

Convert between two angle representations.

angleRatio :: Angle a => a -> a -> Double

Calculate ratio between two angles

Directions in 3D

class Direction d where

Direction is a type class representing directions in R3. The interface is based on that of the Angle class in 2D.

Methods

toSpherical :: Angle a => d -> Spherical a

Convert to polar angles

fromSpherical :: Angle a => Spherical a -> d

Convert from polar angles

data Spherical a

A direction expressed as a pair of spherical coordinates. `Spherical 0 0` is the direction of unitX. The first coordinate represents rotation about the Z axis, the second rotation towards the Z axis.

Constructors

Spherical a a

asSpherical :: Spherical Turn -> Spherical Turn

The identity function with a restricted type, for conveniently restricting unwanted polymorphism. For example, fromDirection . asSpherical . camForward gives a unit vector pointing in the direction of the camera view. Without asSpherical, the intermediate type would be ambiguous.