Diagrams.Trace

Description

"Traces", aka embedded raytracers, for finding points on the edge of a diagram. See Diagrams.Core.Trace for internal implementation details.

Types

data Trace v

class (Ord (Scalar (V a)), VectorSpace (V a)) => Traced a

Diagram traces

trace :: (InnerSpace v, HasLinearMap v, OrderedField (Scalar v), Semigroup m) => Lens' (QDiagram b v m) (Trace v)

setTrace :: (OrderedField (Scalar v), InnerSpace v, HasLinearMap v, Semigroup m) => Trace v -> QDiagram b v m -> QDiagram b v m

withTrace :: (HasLinearMap (V a), Traced a, OrderedField (Scalar (V a)), InnerSpace (V a), Monoid' m) => a -> QDiagram b (V a) m -> QDiagram b (V a) m

Use the trace from some object as the trace for a diagram, in place of the diagram's default trace.

Querying traces

traceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

traceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

maxTraceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

maxTraceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

Subdiagram traces

boundaryFrom :: (HasLinearMap v, OrderedField (Scalar v), InnerSpace v, Semigroup m) => Subdiagram b v m -> v -> Point v

Compute the furthest point on the boundary of a subdiagram, beginning from the location (local origin) of the subdiagram and moving in the direction of the given vector. If there is no such point, the origin is returned; see also boundaryFromMay.

boundaryFromMay :: (HasLinearMap v, OrderedField (Scalar v), Semigroup m, InnerSpace v) => Subdiagram b v m -> v -> Maybe (Point v)

Compute the furthest point on the boundary of a subdiagram, beginning from the location (local origin) of the subdiagram and moving in the direction of the given vector, or Nothing if there is no such point.