Diagrams.Transform

Description

Affine transformations, parameterized by any vector space. For transformations on particular vector spaces, see e.g. Diagrams.TwoD.Transform.

Transformations

data Transformation v

inv :: HasLinearMap v => Transformation v -> Transformation v

transl :: Transformation v -> v

apply :: HasLinearMap v => Transformation v -> v -> v

papply :: HasLinearMap v => Transformation v -> Point v -> Point v

The Transformable class

class HasLinearMap (V t) => Transformable t where

Methods

transform :: Transformation (V t) -> t -> t

Some specific transformations

translation :: HasLinearMap v => v -> Transformation v

translate :: (Transformable t, HasLinearMap (V t)) => V t -> t -> t

moveTo :: HasOrigin t => Point (V t) -> t -> t

place :: HasOrigin t => t -> Point (V t) -> t

scaling :: (HasLinearMap v, Fractional (Scalar v)) => Scalar v -> Transformation v

scale :: (Transformable t, Fractional (Scalar (V t)), Eq (Scalar (V t))) => Scalar (V t) -> t -> t

Miscellaneous transformation-related utilities

conjugate :: HasLinearMap v => Transformation v -> Transformation v -> Transformation v

Conjugate one transformation by another. conjugate t1 t2 is the transformation which performs first t1, then t2, then the inverse of t1.

under :: Transformable a => (a -> a) -> Transformation (V a) -> a -> a

Carry out some transformation "under" another one: f `under` t first applies t, then f, then the inverse of t. For example, scaleX 2 `under` rotationBy (-1/8 :: Turn) is the transformation which scales by a factor of 2 along the diagonal line y = x.

Note that

   (transform t2) under t1 == transform (conjugate t1 t2)

for all transformations t1 and t2.

The HasOrigin class

class VectorSpace (V t) => HasOrigin t where

Methods

moveOriginTo :: Point (V t) -> t -> t

moveOriginBy :: HasOrigin t => V t -> t -> t