Diagrams.TwoD.Ellipse

Description

Two-dimensional ellipses (and, as a special case, circles).

Ellipse and circle diagrams

unitCircle :: (TrailLike t, V t ~ V2, N t ~ n) => t

A circle of radius 1, with center at the origin.

circle :: (TrailLike t, V t ~ V2, N t ~ n, Transformable t) => n -> t

A circle of the given radius, centered at the origin. As a path, it begins at (r,0).

ellipse :: (TrailLike t, V t ~ V2, N t ~ n, Transformable t) => n -> t

ellipse e constructs an ellipse with eccentricity e by scaling the unit circle in the X direction. The eccentricity must be within the interval [0,1).

ellipseXY :: (TrailLike t, V t ~ V2, N t ~ n, Transformable t) => n -> n -> t

ellipseXY x y creates an axis-aligned ellipse, centered at the origin, with radius x along the x-axis and radius y along the y-axis.