Diagrams.TwoD.Model

Description

Tools for visualizing diagrams' internal model: local origins, envelopes, etc.

Showing the local origin

showOrigin :: (Renderable (Path R2) b, Backend b R2, Monoid' m) => QDiagram b R2 m -> QDiagram b R2 m

Mark the origin of a diagram by placing a red dot 1/50th its size.

showOrigin' :: (Renderable (Path R2) b, Backend b R2, Monoid' m) => OriginOpts -> QDiagram b R2 m -> QDiagram b R2 m

Mark the origin of a diagram, with control over colour and scale of marker dot.

data OriginOpts

Constructors

OriginOpts

Fields

_oColor :: Colour Double

_oScale :: Double

_oMinSize :: Double

oColor :: Lens' OriginOpts (Colour Double)

oScale :: Lens' OriginOpts Double

oMinSize :: Lens' OriginOpts Double

showLabels :: (Renderable Text b, Backend b R2, Semigroup m) => QDiagram b R2 m -> QDiagram b R2 Any