Diagrams.TwoD.Segment.Bernstein

Description

Bernstein polynomials, used internally by code to find intersections of paths. This module is probably not of any relevance to most users of diagrams.

Documentation

data BernsteinPoly n

Constructors

BernsteinPoly

Fields

bernsteinDegree :: Int

bernsteinCoeffs :: [n]

listToBernstein :: Fractional n => [n] -> BernsteinPoly n

Create a bernstein polynomial from a list of coëfficients.

evaluateBernstein :: Fractional n => BernsteinPoly n -> n -> n

Evaluate the bernstein polynomial.

degreeElevate :: Fractional n => BernsteinPoly n -> Int -> BernsteinPoly n

Degree elevate a bernstein polynomial a number of times.

bernsteinDeriv :: Fractional n => BernsteinPoly n -> BernsteinPoly n

Find the derivative of a bernstein polynomial.

evaluateBernsteinDerivs :: Fractional n => BernsteinPoly n -> n -> [n]

Evaluate the bernstein polynomial and its derivatives.