Diagrams.TwoD.Text

Description

Very basic text primitives along with associated attributes.

Creating text diagrams

data Text n

A Text primitive consists of the string contents, text alignment and the transformation to be applied. The transformation is scale invarient, the average scale of the transform should always be 1. All text scaling is obtained from the FontSize attribute.

This constructor should not be used directly. Use text, alignedText or baselineText.

Constructors

Text (T2 n) (TextAlignment n) String

data TextAlignment n

TextAlignment specifies the alignment of the text's origin.

Constructors

BaselineText

BoxAlignedText n n

text :: (TypeableFloat n, Renderable (Text n) b) => String -> QDiagram b V2 n Any

Create a primitive text diagram from the given string, with center alignment, equivalent to alignedText 0.5 0.5.

Note that it takes up no space, as text size information is not available.

topLeftText :: (TypeableFloat n, Renderable (Text n) b) => String -> QDiagram b V2 n Any

Create a primitive text diagram from the given string, origin at the top left corner of the text's bounding box, equivalent to alignedText 0 1.

Note that it takes up no space.

alignedText :: (TypeableFloat n, Renderable (Text n) b) => n -> n -> String -> QDiagram b V2 n Any

Create a primitive text diagram from the given string, with the origin set to a point interpolated within the bounding box. The first parameter varies from 0 (left) to 1 (right), and the second parameter from 0 (bottom) to 1 (top). Some backends do not implement this and instead snap to closest corner or the center.

The height of this box is determined by the font's potential ascent and descent, rather than the height of the particular string.

Note that it takes up no space.

baselineText :: (TypeableFloat n, Renderable (Text n) b) => String -> QDiagram b V2 n Any

Create a primitive text diagram from the given string, with the origin set to be on the baseline, at the beginning (although not bounding). This is the reference point of showText in the Cairo graphics library.

Note that it takes up no space.

mkText :: (TypeableFloat n, Renderable (Text n) b) => TextAlignment n -> String -> QDiagram b V2 n Any

Make a text from a TextAlignment, recommending a fill colour of black and fontSize of local 1.

mkText' :: (TypeableFloat n, Renderable (Text n) b) => TextAlignment n -> String -> QDiagram b V2 n Any

Make a text from a TextAlignment without any default size or fill colour. This is useful is you want to recommend your own using recommendFillColor or recommendFontSize.

Text attributes

Font family

newtype Font

The Font attribute specifies the name of a font family. Inner Font attributes override outer ones.

Constructors

Font (Last String)

_Font :: Iso' Font String

getFont :: Font -> String

Extract the font family name from a Font attribute.

font :: HasStyle a => String -> a -> a

Specify a font family to be used for all text within a diagram.

_font :: Lens' (Style v n) (Maybe String)

Lens onto the font name of a style.

Font size

newtype FontSize n

The FontSize attribute specifies the size of a font's em-square. Inner FontSize attributes override outer ones.

Constructors

FontSize (Recommend (Last n))

_FontSize :: Iso' (FontSize n) (Recommend n)

fontSize :: (N a ~ n, Typeable n, HasStyle a) => Measure n -> a -> a

Set the font size, that is, the size of the font's em-square as measured within the current local vector space. The default size is local 1 (which is applied by recommendFontSize).

recommendFontSize :: (N a ~ n, Typeable n, HasStyle a) => Measure n -> a -> a

Recommend a font size. Any use of fontSize above this will overwrite any recommended size. This should only be used with mkText', other text functions already has a recommended font size so this will be ignored.

fontSizeN :: (N a ~ n, Typeable n, Num n, HasStyle a) => n -> a -> a

A convenient synonym for 'fontSize (Normalized w)'.

fontSizeO :: (N a ~ n, Typeable n, HasStyle a) => n -> a -> a

A convenient synonym for 'fontSize (Output w)'.

fontSizeL :: (N a ~ n, Typeable n, Num n, HasStyle a) => n -> a -> a

A convenient sysnonym for 'fontSize (Local w)'.

fontSizeG :: (N a ~ n, Typeable n, Num n, HasStyle a) => n -> a -> a

A convenient synonym for 'fontSize (Global w)'.

getFontSize :: FontSize n -> n

Extract the size from a FontSize attribute.

fontSizeM :: (N a ~ n, Typeable n, HasStyle a) => FontSizeM n -> a -> a

Apply a FontSize attribute.

_fontSizeR :: (Typeable n, OrderedField n) => Lens' (Style v n) (Measured n (Recommend n))

_fontSize :: (Typeable n, OrderedField n) => Lens' (Style v n) (Measure n)

Lens to commit a font size. This is *not* a valid lens (see commited.

_fontSizeU :: Typeable n => Lens' (Style v n) (Maybe n)

Font slant

data FontSlant

The FontSlantA attribute specifies the slant (normal, italic, or oblique) that should be used for all text within a diagram. Inner FontSlantA attributes override outer ones.

Constructors

FontSlantNormal

FontSlantItalic

FontSlantOblique

getFontSlant :: FontSlant -> FontSlant

Extract the font slant from a FontSlantA attribute.

fontSlant :: HasStyle a => FontSlant -> a -> a

Specify the slant (normal, italic, or oblique) that should be used for all text within a diagram. See also italic and oblique for useful special cases.

italic :: HasStyle a => a -> a

Set all text in italics.

oblique :: HasStyle a => a -> a

Set all text using an oblique slant.

_fontSlant :: Lens' (Style v n) FontSlant

Lens onto the font slant in a style.

Font weight

data FontWeight

The FontWeightA attribute specifies the weight (normal or bold) that should be used for all text within a diagram. Inner FontWeightA attributes override outer ones.

Constructors

FontWeightNormal

FontWeightBold

FontWeightBolder

FontWeightLighter

FontWeightThin

FontWeightUltraLight

FontWeightLight

FontWeightMedium

FontWeightSemiBold

FontWeightUltraBold

FontWeightHeavy

getFontWeight :: FontWeight -> FontWeight

Extract the font weight.

fontWeight :: HasStyle a => FontWeight -> a -> a

Specify the weight (normal, bolder, lighter or bold) that should be used for all text within a diagram. See also bold for a useful special case.

bold :: HasStyle a => a -> a

Set all text using a bold font weight.

bolder :: HasStyle a => a -> a

Set all text to be bolder than the inherited font weight.

lighter :: HasStyle a => a -> a

Set all text to be lighter than the inherited font weight.

_fontWeight :: Lens' (Style v n) FontWeight

Lens onto the font weight in a style.

thinWeight :: HasStyle a => a -> a

Set all text using a thin font weight.

ultraLight :: HasStyle a => a -> a

Set all text using a extra light font weight.

light :: HasStyle a => a -> a

Set all text using a light font weight.

mediumWeight :: HasStyle a => a -> a

Set all text using a medium font weight.

heavy :: HasStyle a => a -> a

Set all text using a heavy/black font weight.

semiBold :: HasStyle a => a -> a

Set all text using a semi-bold font weight.

ultraBold :: HasStyle a => a -> a

Set all text using an ultra-bold font weight.