Diagrams.TwoD.Types

Description

Basic types for two-dimensional Euclidean space.

2D Euclidean space

data R2

The two-dimensional Euclidean vector space R^2. This type is intentionally abstract.

To construct a vector, use r2, or & (from Diagrams.Coordinates):

 r2 (3,4) :: R2
 3 & 4    :: R2

To construct the vector from the origin to a point p, use p .-. origin.

To convert a vector v into the point obtained by following v from the origin, use origin .+^ v.

To convert a vector back into a pair of components, use unv2 or coords (from Diagrams.Coordinates). These are typically used in conjunction with the ViewPatterns extension:

 foo (unr2 -> (x,y)) = ...
 foo (coords -> x :& y) = ...

r2 :: (Double, Double) -> R2

Construct a 2D vector from a pair of components. See also &.

unr2 :: R2 -> (Double, Double)

Convert a 2D vector back into a pair of components. See also coords.

type P2 = Point R2

Points in R^2. This type is intentionally abstract.

To construct a point, use p2, or & (see Diagrams.Coordinates):

 p2 (3,4)  :: P2
 3 & 4     :: P2

To construct a point from a vector v, use origin .+^ v.

To convert a point p into the vector from the origin to p, use p .-. origin.

To convert a point back into a pair of coordinates, use unp2, or coords (from Diagrams.Coordinates). It's common to use these in conjunction with the ViewPatterns extension:

 foo (unp2 -> (x,y)) = ...
 foo (coords -> x :& y) = ...

p2 :: (Double, Double) -> P2

Construct a 2D point from a pair of coordinates. See also &.

unp2 :: P2 -> (Double, Double)

Convert a 2D point back into a pair of coordinates. See also coords.

type T2 = Transformation R2

Transformations in R^2.

Angles

class Num a => Angle a where

Type class for types that measure angles.

Methods

toCircleFrac :: a -> CircleFrac

Convert to a fraction of a circle.

fromCircleFrac :: CircleFrac -> a

Convert from a fraction of a circle.

newtype CircleFrac

Newtype wrapper used to represent angles as fractions of a circle. For example, 1/3 = tau/3 radians = 120 degrees.

Constructors

CircleFrac

Fields

getCircleFrac :: Double

newtype Rad

Newtype wrapper for representing angles in radians.

Constructors

Rad

Fields

getRad :: Double

newtype Deg

Newtype wrapper for representing angles in degrees.

Constructors

Deg

Fields

getDeg :: Double

fullCircle :: Angle a => a

An angle representing a full circle.

convertAngle :: (Angle a, Angle b) => a -> b

Convert between two angle representations.